Small Fixes
Fixed colour and exact particle mesh sizing.

I spent all day pursuing failed ways of making the particle spawn line up better with the cutoff. The limitations of the particle system are rather frustrating.

I think what I need to do is change how the cutoff works to accomodate the limitations of the particle system, rather than try to make the particle system do something it can't. The particle spray looks okay on its own - the problem is the disconnect between the two. I think it will be easier to make the sprite conform to the particle logic than the reverse.